… src/librustc_incremental/persist/directory.rs
… src/librustc_metadata/decoder.rs