@michaelwoerister 2017-05-31
Some numbers for compiling `ripgrep`:

### New implementation
* initial runs = 24.81s 24.85s 24.72s
* re-runs      = 4.97s 4.91s 4.88s
* rg-depgraph = 6.0 MB
* grep-depgraph = 2.2 MB

### Current implementation
* initial run = 24.82s 24.98s 24.88s
* re-run      = 4.98s 4.87s 4.91s
* rg-depgraph = 4.0 MB
* grep-depgraph = 1.1 MB

I'd say the timings are pretty much equivalent. The on-disk graph gets bigger though. I'll look into a more efficient on-disk representation (which should be pretty similar to what we soon want to do anyway).