@michaelwoerister 2017-06-01
One more thing: This change slightly regresses the output of `Zincremental-info` because it will show `incremental: module XYZ is dirty because DepNode::Hir(<some opaque hash value>) changed or was removed` in the case of an item having been removed. For things that just have changed, you still get to see the `DefPath` and for removed things you at least get to see the *kind* of `DepNode`.

This could be fixed by storing some additional information with the DepGraph in case `-Zincremental-info` or `-Zquery-dep-graph` is passed. But I don't think it's an urgent problem.