@michaelwoerister 2017-06-01
OK, I have a proof of concept implementation of a new dep-graph encoding:

### New DepGraph Encoding
* initial runs = 24.55s 24.75s 24.76s
* re-runs = 4.68s 4.67s 4.71s
* rg-depgraph = 2.4 MB
* grep-depgraph = 1.1 MB

Looks like that slightly improves re-build times and uses less space than before the PR. I'll clean it up and push it.